V0.9.5.1 Can synchronize table witch is indexed by updatable mode. Fix a bug of contains when the words inputed is multiply. Fix a bug of match when the result count large then max return count. Fix a bug of SqlClient when the parameter inputed is the part of one of the other parameters. Fix a bug of QueryCache that it is used static MD5 provider before.

+++ C#/src/Hubble.Data/Hubble.Core/SFQL/Parse/ParseWhere.cs
@@ -1119,7 +1119,7 @@ namespace Hubble.Core.SFQL.Parse
                                         {
                                             int docid = int.Parse(expression.Right[0].Text);
 
-                                            if (_DBProvider.GetDocIdReplaceFieldValue(docid) != int.MaxValue &&
+                                            if (_DBProvider.GetDocIdReplaceFieldValue(docid) != long.MaxValue &&
                                                 !_DBProvider.DelProvider.DocIdDeleted(docid))
                                             {
                                                 Query.DocumentResultForSort[] dresult = new Hubble.Core.Query.DocumentResultForSort[1];
